Make exceptions control flow and functions optional
