+++ src/library.h
@@ -45,8 +45,10 @@ int lammps_get_natoms(void *);
 void lammps_gather_atoms(void *, char *, int, int, void *);
 void lammps_scatter_atoms(void *, char *, int, int, void *);
 
+#ifdef LAMMPS_EXCEPTION
 int lammps_has_error(void *);
 int lammps_get_last_error_message(void *, char *, int);
+#endif
 
 #ifdef __cplusplus
 }
